Changed handling of related posts on gamepage.
Behaviour has only changed minimally: if there are no related posts
then the horizontal rule and the header "Related posts" should
now no longer show.

However, in the background, the way related posts are handled
is different -- all related posts are now added to an array,
so that the looping and checking of tags against posts now only
need happen once, but the information about all the related posts
can be used multiple times without extra cost.

This is what has enabled the hiding of the header.